Moved and renamed old autonomous OpModes to a new package, and renamed the pedro pathing atounomous OpModes to be easier to understand.
