+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RedCloseNineBall.java
@@ -19,7 +19,7 @@ import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 import org.firstinspires.ftc.vision.apriltag.AprilTagDetection;
 
 @Autonomous
-public class BluePedroPathingLong extends OpMode {
+public class RedCloseNineBall extends OpMode {
     MecanumDrive drive = new MecanumDrive();
     Intake intake = new Intake();
     Launcher launcher = new Launcher();
@@ -30,13 +30,17 @@ public class BluePedroPathingLong extends OpMode {
     private Follower follower;
 
 
-    private final Pose startPose = new Pose(55, 8, Math.toRadians(90)); // Start Pose of our robot.
-    private final Pose launchingPose = new Pose(60, 21.5, Math.toRadians(112)); // Where our robot launches from
-    private final Pose pickup1_1Pose = new Pose(40, 8, Math.toRadians(180)); // Ready to pick up balls
-    private final Pose pickup1_2Pose = new Pose(20, 8, Math.toRadians(180)); // Pickup balls away from wall
-    private final Pose pickup1_3Pose = new Pose(10, 8, Math.toRadians(180)); // pickup balls close to wall
+    private final Pose startPose = new Pose(124.6829268292683, 122.73170731707317, Math.toRadians(36)); // Start Pose of our robot.
+    private final Pose launchingPose = new Pose(92, 92, Math.toRadians(45)); // Where our robot launches from
+    private final Pose pickupReady1Pose = new Pose(98, 84, Math.toRadians(0)); // Ready to pick up closest row of balls
+    private final Pose pickup1Pose = new Pose(125, 84, Math.toRadians(0)); // Pick up closest row of balls
+    private final Pose pickupReady2Pose = new Pose(98, 60, Math.toRadians(0)); //Ready to pick up middle row of balls
+    private final Pose pickup2Pose = new Pose(125, 60, Math.toRadians(0)); //Pick up middle row of balls
+    private final Pose pickupReady3 = new Pose(98, 38.5, Math.toRadians(0)); //Ready to pick up far balls
+    private final Pose endPose = new Pose(125, 38.5, Math.toRadians(0)); //Finish with 3 balls
+
     private Path startToLaunching;
-    private PathChain launchingToPickupReady1, pickup1ToPickup2, pickup2ToPickup3, pickup3ToPickup2, pickup3ToLaunching, launchingToPickup2;
+    private PathChain launchingToPickupReady1, pickupReady1ToPickup1, pickup1ToLaunching, launchingToPickupReady2, pickupReady2ToPickup2, pickup2ToLaunching2, launchingToPickupReady3, pickupReady3ToFinish;
 
     public void buildPaths() {
 
@@ -46,32 +50,43 @@ public class BluePedroPathingLong extends OpMode {
 
         //these next several section set up the rest of the paths, created in the PathChain
         launchingToPickupReady1 = follower.pathBuilder()
-                .addPath(new BezierLine(launchingPose, pickup1_1Pose))
-                .setLinearHeadingInterpolation(launchingPose.getHeading(), pickup1_1Pose.getHeading())
+                .addPath(new BezierLine(launchingPose, pickupReady1Pose))
+                .setLinearHeadingInterpolation(launchingPose.getHeading(), pickupReady1Pose.getHeading())
                 .build();
 
-        pickup1ToPickup2 = follower.pathBuilder()
-                .addPath(new BezierLine(pickup1_1Pose, pickup1_2Pose))
-                .setLinearHeadingInterpolation(pickup1_1Pose.getHeading(), pickup1_2Pose.getHeading())
+        pickupReady1ToPickup1 = follower.pathBuilder()
+                .addPath(new BezierLine(pickupReady1Pose, pickup1Pose))
+                .setLinearHeadingInterpolation(pickupReady1Pose.getHeading(), pickup1Pose.getHeading())
                 .build();
 
-        pickup2ToPickup3 = follower.pathBuilder()
-                .addPath(new BezierLine(pickup1_2Pose, pickup1_3Pose))
-                .setLinearHeadingInterpolation(pickup1_2Pose.getHeading(), pickup1_3Pose.getHeading())
+        pickup1ToLaunching = follower.pathBuilder()
+                .addPath(new BezierLine(pickup1Pose, launchingPose))
+                .setLinearHeadingInterpolation(pickup1Pose.getHeading(), launchingPose.getHeading())
                 .build();
 
-        pickup3ToPickup2 = follower.pathBuilder()
-                .addPath(new BezierLine(pickup1_3Pose, pickup1_2Pose))
-                .setLinearHeadingInterpolation(pickup1_3Pose.getHeading(), pickup1_2Pose.getHeading())
+        launchingToPickupReady2 = follower.pathBuilder()
+                .addPath(new BezierLine(launchingPose, pickupReady2Pose))
+                .setLinearHeadingInterpolation(launchingPose.getHeading(), pickupReady2Pose.getHeading())
                 .build();
 
-        pickup3ToLaunching = follower.pathBuilder()
-                .addPath(new BezierLine(pickup1_3Pose, launchingPose))
-                .setLinearHeadingInterpolation(pickup1_3Pose.getHeading(), launchingPose.getHeading())
+        pickupReady2ToPickup2 = follower.pathBuilder()
+                .addPath(new BezierLine(pickupReady2Pose, pickup2Pose))
+                .setLinearHeadingInterpolation(pickupReady2Pose.getHeading(), pickup2Pose.getHeading())
                 .build();
-        launchingToPickup2 = follower.pathBuilder()
-                .addPath(new BezierLine(launchingPose, pickup1_2Pose))
-                .setLinearHeadingInterpolation( launchingPose.getHeading(), pickup1_2Pose.getHeading())
+
+        pickup2ToLaunching2 = follower.pathBuilder()
+                .addPath(new BezierLine(pickup2Pose, launchingPose))
+                .setLinearHeadingInterpolation(pickup2Pose.getHeading(), launchingPose.getHeading())
+                .build();
+
+        launchingToPickupReady3 =  follower.pathBuilder()
+                .addPath(new BezierLine(launchingPose, pickupReady3))
+                .setLinearHeadingInterpolation(launchingPose.getHeading(), pickupReady3.getHeading())
+                .build();
+
+        pickupReady3ToFinish = follower.pathBuilder()
+                .addPath(new BezierLine(pickupReady3, endPose))
+                .setLinearHeadingInterpolation(pickupReady3.getHeading(), endPose.getHeading())
                 .build();
     }
 
@@ -81,35 +96,26 @@ public class BluePedroPathingLong extends OpMode {
         FIND_TAG_1,
         SPIN_UP_1,
         LAUNCHING_1,
-        INTAKE_POSE_1_1,
-        INTAKE_1_2,
-        INTAKE_1_3,
-        INTAKE_1_4,
-        INTAKE_1_5,
-        INTAKE_1_6,
-        INTAKE_1_7,
+        PREPARE_TO_INTAKE_POSE_1,
+        INTAKE_1,
         GO_TO_LAUNCH_2,
         WAIT_TO_FINISH_PATH_2,
         FIND_TAG_2,
         SPIN_UP_2,
         LAUNCHING_2,
-        INTAKE_POSE_2_1,
-        INTAKE_2_2,
-        INTAKE_2_3,
-        INTAKE_2_4,
-        INTAKE_2_5,
-        INTAKE_2_6,
-        INTAKE_2_7,
+        PREPARE_TO_INTAKE_POSE_2,
+        INTAKE_2,
         GO_TO_LAUNCH_3,
         WAIT_TO_FINISH_PATH_3,
         FIND_TAG_3,
         SPIN_UP_3,
         LAUNCHING_3,
+        PREPARE_TO_INTAKE_POSE_3,
         GO_TO_END_POSE,
         FINISHED
     }
 
-    BluePedroPathingLong.State state;
+    RedCloseNineBall.State state;
     ElapsedTime driveTimer = new ElapsedTime();
 
 
@@ -152,11 +158,11 @@ public class BluePedroPathingLong extends OpMode {
                 state == State.LAUNCHING_2 ||
                 state == State.FIND_TAG_3 ||
                 state == State.SPIN_UP_3 ||
-                state == State.LAUNCHING_3 )
+                state == State.LAUNCHING_3)
         {
             doAprilTag();
         }
-        AprilTagDetection id24 = aprilTagWebcam.getTagBySpecificId(20);
+        AprilTagDetection id24 = aprilTagWebcam.getTagBySpecificId(24);
 
         switch (state) {
             case GO_TO_LAUNCH_1:
@@ -182,7 +188,7 @@ public class BluePedroPathingLong extends OpMode {
                 }
                 break;
             case LAUNCHING_1:
-                if (driveTimer.seconds() < 2) {
+                if (driveTimer.seconds() < 3) {
                     intake.startIntake();
                     launcher.loadBall();
                 }
@@ -192,62 +198,26 @@ public class BluePedroPathingLong extends OpMode {
                     launcher.stopLauncher();
                     launcher.resetFeeder();
                     //Launcher.LaunchState = Launcher.LaunchState.IDLE;
-                    state = State.INTAKE_POSE_1_1;
+                    state = State.PREPARE_TO_INTAKE_POSE_1;
                     driveTimer.reset();
                 }
                 break;
-            case INTAKE_POSE_1_1:
+            case PREPARE_TO_INTAKE_POSE_1:
                 if(!follower.isBusy()){
-                    intake.startIntake();
                     follower.followPath(launchingToPickupReady1, true);
-                    state = State.INTAKE_1_2;
+                    state = State.INTAKE_1;
                 }
                 break;
-            case INTAKE_1_2:
+            case INTAKE_1:
                 if(!follower.isBusy()){
-                    follower.followPath(pickup1ToPickup2, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_1_3;
-                }
-                break;
-            case INTAKE_1_3:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup2ToPickup3, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_1_4;
-                }
-                break;
-            case INTAKE_1_4:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup3ToPickup2, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_1_5;
-                }
-                break;
-            case INTAKE_1_5:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup2ToPickup3, .7, false);
-                    intake.startIntake();
-                    state = State.GO_TO_LAUNCH_2;
-                }
-                break;
-            case INTAKE_1_6:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup3ToPickup2, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_1_7;
-                }
-                break;
-            case INTAKE_1_7:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup2ToPickup3, .7, false);
+                    follower.followPath(pickupReady1ToPickup1, .4, false);
                     intake.startIntake();
                     state = State.GO_TO_LAUNCH_2;
                 }
                 break;
             case GO_TO_LAUNCH_2:
                 if(!follower.isBusy()){
-                    follower.followPath(pickup3ToLaunching);
+                    follower.followPath(pickup1ToLaunching);
                     state = State.WAIT_TO_FINISH_PATH_2;
                 }
                 break;
@@ -280,62 +250,26 @@ public class BluePedroPathingLong extends OpMode {
                     launcher.resetFeeder();
                     Launcher.LaunchState = Launcher.LaunchState.IDLE;
                     launcher.stopLauncher();
-                    state = State.INTAKE_POSE_2_1;
+                    state = State.PREPARE_TO_INTAKE_POSE_2;
                     driveTimer.reset();
                 }
                 break;
-            case INTAKE_POSE_2_1:
+            case PREPARE_TO_INTAKE_POSE_2:
                 if(!follower.isBusy()){
-                    intake.startIntake();
-                    follower.followPath(launchingToPickupReady1, true);
-                    state = State.INTAKE_2_2;
+                    follower.followPath(launchingToPickupReady2, true);
+                    state = State.INTAKE_2;
                 }
                 break;
-            case INTAKE_2_2:
+            case INTAKE_2:
                 if(!follower.isBusy()){
-                    follower.followPath(pickup1ToPickup2, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_2_3;
-                }
-                break;
-            case INTAKE_2_3:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup2ToPickup3, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_2_4;
-                }
-                break;
-            case INTAKE_2_4:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup3ToPickup2, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_2_5;
-                }
-                break;
-            case INTAKE_2_5:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup2ToPickup3, .7, false);
-                    intake.startIntake();
-                    state = State.GO_TO_LAUNCH_3;
-                }
-                break;
-            case INTAKE_2_6:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup3ToPickup2, .7, false);
-                    intake.startIntake();
-                    state = State.INTAKE_2_7;
-                }
-                break;
-            case INTAKE_2_7:
-                if(!follower.isBusy()){
-                    follower.followPath(pickup2ToPickup3, .7, false);
+                    follower.followPath(pickupReady2ToPickup2, .4, false);
                     intake.startIntake();
                     state = State.GO_TO_LAUNCH_3;
                 }
                 break;
             case GO_TO_LAUNCH_3:
                 if(!follower.isBusy()){
-                    follower.followPath(pickup3ToLaunching);
+                    follower.followPath(pickup2ToLaunching2);
                     state = State.WAIT_TO_FINISH_PATH_3;
                 }
                 break;
@@ -368,21 +302,27 @@ public class BluePedroPathingLong extends OpMode {
                     launcher.resetFeeder();
                     Launcher.LaunchState = Launcher.LaunchState.IDLE;
                     launcher.stopLauncher();
-                    state = State.GO_TO_END_POSE;
+                    state = State.PREPARE_TO_INTAKE_POSE_3;
                     driveTimer.reset();
                 }
                 break;
+            case PREPARE_TO_INTAKE_POSE_3:
+                if(!follower.isBusy()){
+                    follower.followPath(launchingToPickupReady3, true);
+                    state = State.GO_TO_END_POSE;
+                }
+
             case GO_TO_END_POSE:
                 if(!follower.isBusy()){
                     intake.startIntake();
-                    follower.followPath(launchingToPickup2);
+                    follower.followPath(pickupReady3ToFinish);
                     driveTimer.reset();
                     state = State.FINISHED;
                 }
                 break;
             case FINISHED:
                 if(!follower.isBusy()) {
-                    if(driveTimer.seconds() > 2)
+                        if(driveTimer.seconds() > 2)
                         intake.stopIntake();
                 }
                 break;
@@ -393,13 +333,13 @@ public class BluePedroPathingLong extends OpMode {
         }
 
 
-    }
+        }
 
 
     private void doAprilTag() {
         //Update the vision portal
         aprilTagWebcam.update();
-        AprilTagDetection id24 = aprilTagWebcam.getTagBySpecificId(20); // TAG ID 24 is the red goal
+        AprilTagDetection id24 = aprilTagWebcam.getTagBySpecificId(24); // TAG ID 24 is the red goal
         //aprilTagWebcam.displayDetectionTelemetry(id24);
         // NOTE: we will need a separate OPMODE (otherwise identical) that sets the target TAGID to BLUE (#20)
         if (id24 != null && id24.ftcPose != null) {
